requirements.txt
Django>=4.2,<5.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
paho-mqtt>=2.0.0
python-dotenv==1.0.0
whitenoise==6.6.0
weasyprint>=60.0
qrcode[pil]>=7.0
python-barcode>=0.15.1
